+++ subsys/net/l2/wifi/wifi_mgmt.c
@@ -371,7 +371,7 @@ static const struct wifi_mgmt_ops *const get_wifi_api(struct net_if *iface)
 	return off_api ? off_api->wifi_mgmt_api : NULL;
 }
 
-static int wifi_connect(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_connect(uint32_t mgmt_request, struct net_if *iface,
 			void *data, size_t len)
 {
 	struct wifi_connect_req_params *params =
@@ -475,7 +475,7 @@ static void scan_result_cb(struct net_if *iface, int status,
 #endif /* CONFIG_WIFI_MGMT_RAW_SCAN_RESULTS_ONLY */
 }
 
-static int wifi_scan(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_scan(uint32_t mgmt_request, struct net_if *iface,
 		     void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -504,7 +504,7 @@ static int wifi_scan(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_SCAN, wifi_scan);
 
-static int wifi_disconnect(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_disconnect(uint32_t mgmt_request, struct net_if *iface,
 			   void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -546,7 +546,7 @@ void wifi_mgmt_raise_disconnect_result_event(struct net_if *iface, int status)
 }
 
 #ifdef CONFIG_WIFI_NM_WPA_SUPPLICANT_ROAMING
-static int wifi_start_roaming(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_start_roaming(uint32_t mgmt_request, struct net_if *iface,
 			      void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -589,7 +589,7 @@ static int wifi_start_roaming(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_START_ROAMING, wifi_start_roaming);
 
-static int wifi_neighbor_rep_complete(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_neighbor_rep_complete(uint32_t mgmt_request, struct net_if *iface,
 				      void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -684,7 +684,7 @@ void wifi_mgmt_raise_neighbor_rep_recv_event(struct net_if *iface, char *inbuf, 
 }
 #endif
 
-static int wifi_ap_enable(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_ap_enable(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	struct wifi_connect_req_params *params =
@@ -705,7 +705,7 @@ static int wifi_ap_enable(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_AP_ENABLE, wifi_ap_enable);
 
-static int wifi_ap_disable(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_ap_disable(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -724,7 +724,7 @@ static int wifi_ap_disable(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_AP_DISABLE, wifi_ap_disable);
 
-static int wifi_ap_sta_disconnect(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_ap_sta_disconnect(uint32_t mgmt_request, struct net_if *iface,
 				  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -752,7 +752,7 @@ static int wifi_ap_sta_disconnect(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_AP_STA_DISCONNECT, wifi_ap_sta_disconnect);
 
-static int wifi_ap_config_params(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_ap_config_params(uint32_t mgmt_request, struct net_if *iface,
 				 void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -790,7 +790,7 @@ static int wifi_ap_config_params(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_AP_CONFIG_PARAM, wifi_ap_config_params);
 
-static int wifi_ap_set_rts_threshold(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_ap_set_rts_threshold(uint32_t mgmt_request, struct net_if *iface,
 				  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -814,7 +814,7 @@ static int wifi_ap_set_rts_threshold(uint64_t mgmt_request, struct net_if *iface
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_AP_RTS_THRESHOLD, wifi_ap_set_rts_threshold);
 
-static int wifi_iface_status(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_iface_status(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -842,7 +842,7 @@ void wifi_mgmt_raise_iface_status_event(struct net_if *iface,
 }
 
 #ifdef CONFIG_NET_STATISTICS_WIFI
-static int wifi_iface_stats(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_iface_stats(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -861,7 +861,7 @@ static int wifi_iface_stats(uint64_t mgmt_request, struct net_if *iface,
 }
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_STATS_GET_WIFI, wifi_iface_stats);
 
-static int wifi_iface_stats_reset(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_iface_stats_reset(uint32_t mgmt_request, struct net_if *iface,
 				  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -876,7 +876,7 @@ static int wifi_iface_stats_reset(uint64_t mgmt_request, struct net_if *iface,
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_STATS_RESET_WIFI, wifi_iface_stats_reset);
 #endif /* CONFIG_NET_STATISTICS_WIFI */
 
-static int wifi_11k_cfg(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_11k_cfg(uint32_t mgmt_request, struct net_if *iface,
 			void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -902,7 +902,7 @@ static int wifi_11k_cfg(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_11K_CONFIG, wifi_11k_cfg);
 
-static int wifi_11k_neighbor_request(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_11k_neighbor_request(uint32_t mgmt_request, struct net_if *iface,
 				     void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -923,7 +923,7 @@ static int wifi_11k_neighbor_request(uint64_t mgmt_request, struct net_if *iface
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_11K_NEIGHBOR_REQUEST,
 				  wifi_11k_neighbor_request);
 
-static int wifi_set_power_save(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_set_power_save(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -977,7 +977,7 @@ static int wifi_set_power_save(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_PS, wifi_set_power_save);
 
-static int wifi_get_power_save_config(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_get_power_save_config(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1001,7 +1001,7 @@ static int wifi_get_power_save_config(uint64_t mgmt_request, struct net_if *ifac
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_PS_CONFIG, wifi_get_power_save_config);
 
-static int wifi_set_twt(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_set_twt(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1068,7 +1068,7 @@ fail:
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_TWT, wifi_set_twt);
 
-static int wifi_set_btwt(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_set_btwt(uint32_t mgmt_request, struct net_if *iface,
 			  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1108,7 +1108,7 @@ void wifi_mgmt_raise_twt_event(struct net_if *iface, struct wifi_twt_params *twt
 					sizeof(struct wifi_twt_params));
 }
 
-static int wifi_reg_domain(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_reg_domain(uint32_t mgmt_request, struct net_if *iface,
 			   void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1140,7 +1140,7 @@ void wifi_mgmt_raise_twt_sleep_state(struct net_if *iface,
 					sizeof(twt_sleep_state));
 }
 
-static int wifi_mode(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_mode(uint32_t mgmt_request, struct net_if *iface,
 				void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1164,7 +1164,7 @@ static int wifi_mode(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_MODE, wifi_mode);
 
-static int wifi_packet_filter(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_packet_filter(uint32_t mgmt_request, struct net_if *iface,
 				void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1188,7 +1188,7 @@ static int wifi_packet_filter(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_PACKET_FILTER, wifi_packet_filter);
 
-static int wifi_channel(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_channel(uint32_t mgmt_request, struct net_if *iface,
 				void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1212,7 +1212,7 @@ static int wifi_channel(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_CHANNEL, wifi_channel);
 
-static int wifi_get_version(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_get_version(uint32_t mgmt_request, struct net_if *iface,
 			   void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1228,7 +1228,7 @@ static int wifi_get_version(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_VERSION, wifi_get_version);
 
-static int wifi_btm_query(uint64_t mgmt_request, struct net_if *iface, void *data, size_t len)
+static int wifi_btm_query(uint32_t mgmt_request, struct net_if *iface, void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
 	const struct wifi_mgmt_ops *const wifi_mgmt_api = get_wifi_api(iface);
@@ -1252,7 +1252,7 @@ static int wifi_btm_query(uint64_t mgmt_request, struct net_if *iface, void *dat
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_BTM_QUERY, wifi_btm_query);
 
-static int wifi_get_connection_params(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_get_connection_params(uint32_t mgmt_request, struct net_if *iface,
 			void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1272,7 +1272,7 @@ static int wifi_get_connection_params(uint64_t mgmt_request, struct net_if *ifac
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_CONN_PARAMS, wifi_get_connection_params);
 
-static int wifi_wps_config(uint64_t mgmt_request, struct net_if *iface, void *data, size_t len)
+static int wifi_wps_config(uint32_t mgmt_request, struct net_if *iface, void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
 	const struct wifi_mgmt_ops *const wifi_mgmt_api = get_wifi_api(iface);
@@ -1291,7 +1291,7 @@ static int wifi_wps_config(uint64_t mgmt_request, struct net_if *iface, void *da
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_WPS_CONFIG, wifi_wps_config);
 
-static int wifi_set_rts_threshold(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_set_rts_threshold(uint32_t mgmt_request, struct net_if *iface,
 				  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1316,7 +1316,7 @@ static int wifi_set_rts_threshold(uint64_t mgmt_request, struct net_if *iface,
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_RTS_THRESHOLD, wifi_set_rts_threshold);
 
 #ifdef CONFIG_WIFI_NM_WPA_SUPPLICANT_DPP
-static int wifi_dpp(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_dpp(uint32_t mgmt_request, struct net_if *iface,
 		    void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1338,7 +1338,7 @@ NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_DPP, wifi_dpp);
 
 #endif /* CONFIG_WIFI_NM_WPA_SUPPLICANT_DPP */
 
-static int wifi_pmksa_flush(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_pmksa_flush(uint32_t mgmt_request, struct net_if *iface,
 					   void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1357,7 +1357,7 @@ static int wifi_pmksa_flush(uint64_t mgmt_request, struct net_if *iface,
 
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_PMKSA_FLUSH, wifi_pmksa_flush);
 
-static int wifi_get_rts_threshold(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_get_rts_threshold(uint32_t mgmt_request, struct net_if *iface,
 				  void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1382,7 +1382,7 @@ static int wifi_get_rts_threshold(uint64_t mgmt_request, struct net_if *iface,
 NET_MGMT_REGISTER_REQUEST_HANDLER(NET_REQUEST_WIFI_RTS_THRESHOLD_CONFIG, wifi_get_rts_threshold);
 
 #ifdef CONFIG_WIFI_NM_WPA_SUPPLICANT_CRYPTO_ENTERPRISE
-static int wifi_set_enterprise_creds(uint64_t mgmt_request, struct net_if *iface,
+static int wifi_set_enterprise_creds(uint32_t mgmt_request, struct net_if *iface,
 					   void *data, size_t len)
 {
 	const struct device *dev = net_if_get_device(iface);
@@ -1721,7 +1721,7 @@ static int add_static_network_config(struct net_if *iface)
 #endif /* defined(CONFIG_WIFI_CREDENTIALS_STATIC) */
 }
 
-static int connect_stored_command(uint64_t mgmt_request, struct net_if *iface, void *data,
+static int connect_stored_command(uint32_t mgmt_request, struct net_if *iface, void *data,
 				  size_t len)
 {
 	int ret = 0;
