+++ subsys/net/lib/dhcpv6/dhcpv6.c
@@ -2200,7 +2200,7 @@ static void dhcpv6_timeout(struct k_work *work)
 }
 
 static void dhcpv6_iface_event_handler(struct net_mgmt_event_callback *cb,
-				       uint64_t mgmt_event, struct net_if *iface)
+				       uint32_t mgmt_event, struct net_if *iface)
 {
 	sys_snode_t *node = NULL;
